Disable theme button, increase req limit, refactor
