+++ app/javascript/components/common/ThemeToggleButton.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback, useState } from 'react'
 import { GraphicalIcon } from './GraphicalIcon'
-import { setThemeClassName } from '../settings/ThemePreferenceForm'
+import { setThemeClassName } from '../settings/theme-preference-form/utils'
 
 export function ThemeToggleButton(): JSX.Element {
   const { currentColorScheme, switchToColorMode } = useSwitchTheme()
