+++ app/javascript/components/settings/theme-preference-form/useTheme.tsx
@@ -1,0 +1,41 @@
+import { useCallback, useState } from 'react'
+import { useSettingsMutation } from '../useSettingsMutation'
+import { setThemeClassName } from './utils'
+import { Theme, ThemePreferenceLinks } from '../ThemePreferenceForm'
+import { QueryStatus } from 'react-query'
+
+type RequestBody = {
+  user_preferences: {
+    theme: string
+  }
+}
+
+type useThemeReturns = {
+  handleThemeUpdate: (t: Theme) => void
+  status: QueryStatus
+  error: unknown
+  theme: string
+}
+export function useTheme(
+  defaultThemePreference: string,
+  links: ThemePreferenceLinks
+): useThemeReturns {
+  const [theme, setTheme] = useState<string>(defaultThemePreference || '')
+
+  const { mutation, status, error } = useSettingsMutation<RequestBody>({
+    endpoint: links.update,
+    method: 'PATCH',
+    body: { user_preferences: { theme } },
+  })
+
+  const handleThemeUpdate = useCallback(
+    (t) => {
+      mutation()
+      setTheme(t.value)
+      setThemeClassName(t.value)
+    },
+    [mutation]
+  )
+
+  return { handleThemeUpdate, status, error, theme }
+}
